+++ gatspy/periodic/__init__.py
@@ -6,13 +6,14 @@ from __future__ import absolute_import
 
 __all__ = ['LombScargle', 'LombScargleFast', 'LombScargleAstroML',
            'LombScargleMultiband', 'LombScargleMultibandFast',
-           'SuperSmoother', 'SuperSmootherMultiband',
+           'TrendedLombScargle', 'SuperSmoother', 'SuperSmootherMultiband',
            'RRLyraeTemplateModeler', 'RRLyraeTemplateModelerMultiband',
            'NaiveMultiband']
 
 from .lomb_scargle import *
 from .lomb_scargle_fast import *
 from .lomb_scargle_multiband import *
+from .trended_lomb_scargle import *
 from .supersmoother import *
 from .template_modeler import *
 from .naive_multiband import *