+++ CHANGES.md
@@ -5,6 +5,8 @@
 - Fix bug in ``predict()`` for ``LombScargle`` methods when ``center_data=False``
 - Fix bug in ``score()`` for ``LombScargle`` methods when ``center_data=False`` and ``fit_offset=False``
 - Make ``supersmoother`` a soft dependency
+- New feature: ``gatspy.periodic.TrendedLombScargle`` extends ``LombScargle``
+  by adding a linear trend parameter ``d*t`` to the fitted model
 
 ## (v0.2.1) bugfix release (19 Aug 2015)